# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateProject:
    properties:
      description:
        type: string
      level:
        type: string
      title:
        type: string
      work_hours:
        type: integer
    type: object
...
  dto.UpdateProject:
    properties:
      description:
        type: string
      level:
        type: string
      title:
        type: string
      work_hours:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend for AITU Corporate Self-Study Portal.
  title: ACSP Backend
  version: "1.0"
paths:
...
  /api/v1/coding-lab/disciplines/{id}/projects:
    get:
      consumes:
      - application/json
      description: Get all projects of discipline
      operationId: get-all-projects-by-discipline-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all projects by discipline id
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Method for creating a project for a user by id in the database
        by user id
      operationId: create-project
      parameters:
      - description: project information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProject'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a project
      tags:
      - projects
  /api/v1/coding-lab/disciplines/{id}/projects/{projectID}:
    delete:
      consumes:
      - application/json
      description: Delete a project by id
      operationId: delete-project-by-id
      parameters:
      - description: discipline id
        in: path
        name: id
        required: true
        type: integer
      - description: project id
        in: path
        name: projectID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project deleted
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a project by id
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Get a project by id
      operationId: get-project-by-id
      parameters:
      - description: discipline id
        in: path
        name: id
        required: true
        type: integer
      - description: project id
        in: path
        name: projectID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get a project by id
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update data of project
      operationId: update-project-by-id
      parameters:
      - description: discipline id
        in: path
        name: id
        required: true
        type: integer
      - description: project id
        in: path
        name: projectID
        required: true
        type: integer
      - description: project information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProject'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a project by id
      tags:
      - projects
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"